refactor : 경로 조회시 type을 받도록 변경
- type을 통해 검색 유형이 변경된다
