+++ src/main/java/nextstep/subway/applicaion/PathService.java
@@ -1,10 +1,7 @@
 package nextstep.subway.applicaion;
 
 import nextstep.subway.applicaion.dto.PathResponse;
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.Path;
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.SubwayMap;
+import nextstep.subway.domain.*;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,12 +16,11 @@ public class PathService {
         this.stationService = stationService;
     }
 
-    public PathResponse findPath(Long source, Long target, boolean minimumTime) {
+    public PathResponse findPath(Long source, Long target, SectionPathType type) {
         Station upStation = stationService.findById(source);
         Station downStation = stationService.findById(target);
         List<Line> lines = lineService.findLines();
-        SubwayMap.SectionPathType sectionPathType = minimumTime ? SubwayMap.SectionPathType.DURATION : SubwayMap.SectionPathType.DISTANCE;
-        SubwayMap subwayMap = new SubwayMap(lines, sectionPathType);
+        SubwayMap subwayMap = new SubwayMap(lines, type);
         Path path = subwayMap.findPath(upStation, downStation);
         return PathResponse.of(path);
     }
