+++ src/main/java/nextstep/subway/domain/SubwayMap.java
@@ -73,19 +73,6 @@ public class SubwayMap {
                 });
     }
 
-    public enum SectionPathType {
-        DISTANCE, DURATION;
 
-        public void setEdgeWeight(SimpleDirectedWeightedGraph<Station, SectionEdge> graph, Section it, SectionEdge sectionEdge) {
-            switch (this) {
-                case DISTANCE:
-                    graph.setEdgeWeight(sectionEdge, it.getDistance());
-                    break;
-                case DURATION:
-                    graph.setEdgeWeight(sectionEdge, it.getDuration());
-                    break;
-            }
-        }
-    }
 
 }
